fix: correct indentation in architecture-validator.py
Fixed SyntaxError caused by incorrect indentation on line 111
in the is_generated assignment within try block.

Also fixed category parameter to use 'file_sizes' instead of 'files'.

+++ services/sandevistan-service-go/pkg/api/oas_schemas_gen.go
@@ -177,7 +177,7 @@ func (s *ErrorDetails) init() ErrorDetails {
 }
 
 // NewOptFloat32 returns new OptFloat32 with value set to v.
-func NewOptFloat32(v float32) OptFloat32 {
+func NewOptFloat32(v float64) OptFloat32 {
 	return OptFloat32{
 		Value: v,
 		Set:   true,
